+++ src/components/ui/terminal/renderCommand.tsx
@@ -1,0 +1,16 @@
+import { CommandOutput } from "@/components/ui/terminal/CommandOutput";
+import React from "react";
+import { renderCommandContent } from "@/components/ui/terminal/renderCommandContent";
+import { Command } from "@/app/terminal/types";
+
+
+export const renderCommand = (command: Command): JSX.Element => {
+
+  return (
+    <CommandOutput>
+      {command.content.map((contentItem) => renderCommandContent(contentItem)
+      )}
+    </CommandOutput>
+  );
+
+};
